move magic share struct to the root of the library and make the cli reflect the most recent changes
