+++ lib/Cargo.toml
@@ -10,9 +10,11 @@ adi = { path = "../adi" }
 djb = { path = "../djb" }
 s39 = { path = "../s39" }
 
+argon2 = "0.1"
 base64 = "0.13"
 blake3 = "0.3"
+chrono = "0.4"
 rand = "0.8"
-argon2 = "0.1"
+serde = { version = "1.0", features = ["derive"] }
 strum = "0.21"
 strum_macros = "0.21"
